database: declare config sub-structs, add port as database option
