+++ test/config.json
@@ -2,6 +2,7 @@
 	"database": {
 		"name": "temporal",
 		"url": "127.0.0.1",
+		"port": "5433",
 		"username": "postgres",
 		"password": "password123"
 	},
